requirements.txt
asgiref==3.8.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.7.1
dj-rest-auth==2.2.7
Django==4.2.17
django-allauth==0.54.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-auth==2.37.0
idna==3.10
oauthlib==3.2.2
pillow==11.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
